# specs/001-agent-orchestration/contracts/api.yaml
openapi: 3.0.3
info:
  title: Multi-Agent Code Development Orchestration Platform API
  description: REST API for orchestrating AI agents in code development workflows
  version: 1.0.0
  contact:
    name: API Support
    email: support@agent-orchestration.local
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.agent-orchestration.com/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /workflows:
    get:
      summary: List available workflows
      description: Returns a list of all available workflow templates
      operationId: listWorkflows
      tags:
        - Workflows
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowSummary'
                  total:
                    type: integer
                    minimum: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Execute workflow
      description: Execute a workflow with the provided parameters
      operationId: executeWorkflow
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionRequest'
      responses:
        '202':
          description: Workflow execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'
        '400':
          description: Invalid workflow request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    WorkflowSummary:
      type: object
      required:
        - id
        - name
        - description
        - type
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [simple, langgraph]
        tags:
          type: array
          items:
            type: string
    WorkflowExecutionRequest:
      type: object
      required:
        - workflowId
      properties:
        workflowId:
          type: string
          description: ID of the workflow to execute
        parameters:
          type: object
          description: Workflow-specific parameters
        workspaceDir:
          type: string
          description: Working directory for execution
        timeoutSeconds:
          type: integer
          minimum: 10
          maximum: 3600
          default: 300
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    WorkflowExecutionResponse:
      type: object
      required:
        - executionId
        - status
      properties:
        executionId:
          type: string
          description: Unique identifier for this execution
        status:
          type: string
          enum: [pending, running, completed, failed]
        estimatedDuration:
          type: integer
          description: Estimated completion time in seconds
        message:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT